+++ .github/kubernetes/app/web/hooks/alerts/store-alert.sql
@@ -1,0 +1,3 @@
+insert into probe_alerts (event_id, code, node, scope)
+values (/* eventId */ 'evt', /* code */ 'TQL-OPS-9006', /* node */ 'pod', /* scope */ 'database')
+;